GL14/volume/3D blending updated

+++ assets/gl/gl21/glvolume_3d_frag.glsl
@@ -127,7 +127,7 @@ uniform vec3 rayStep;
 /*
  * Length of the rayStep vector.
  */
-uniform int stepLength;
+uniform float stepLength;
